% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_date.R
\name{check_day_correctly_entered_dmy}
\alias{check_day_correctly_entered_dmy}
\title{Check if Day Component is Valid in dmy date type}
\usage{
check_day_correctly_entered_dmy(data_frame, column_name, separator = "-")
}
\arguments{
\item{data_frame}{data frame}

\item{column_name}{name of column that keeps dates to be imputed}

\item{separator}{by default "-" it is a day-month-year separator, for example "2024-10-21" has "-" separator}
}
\value{
A logical vector
}
\description{
This function checks whether the day component in a vector of date strings is valid,
i.e., not exceeding the maximum number of days for the given month and year.
It returns a logical vector indicating which elements have a correctly specified day.
}
\author{
Lukasz Andrzejewski
}
