% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_date.R
\name{check_if_only_year_entered}
\alias{check_if_only_year_entered}
\title{Check if a Vector Contains only Year}
\usage{
check_if_only_year_entered(
  data_frame,
  column_name,
  separator = "-",
  month = "UNK",
  day = "UN"
)
}
\arguments{
\item{data_frame}{data frame}

\item{column_name}{name of column that keeps dates to be imputed}

\item{separator}{by default "-" it is a day-month-year separator, for example "2024-10-21" has "-" separator}

\item{month}{by default "UNK" - the format of unknown month}

\item{day}{by default "UN" - the format of unknown day}
}
\value{
A logical vector
}
\description{
This function determines whether the elements of a vector contain only **year**
It returns a logical vector indicating which elements meet this criterion.
}
\author{
Lukasz Andrzejewski
}
