% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_date.R
\name{check_if_year_month_day_entered}
\alias{check_if_year_month_day_entered}
\title{Check if a vector contains a complete date}
\usage{
check_if_year_month_day_entered(
  data_frame,
  column_name,
  separator = "-",
  date_format = "ymd",
  year = "UNKN",
  month = "UNK",
  day = "UN"
)
}
\arguments{
\item{data_frame}{data frame}

\item{column_name}{name of column that keeps dates to be imputed}

\item{separator}{by default "-" it is a day-month-year separator, for example "2024-10-21" has "-" separator}

\item{date_format}{by default "ymd". choose between ymd (if first year, then month then day) and dmy (if first day, then month then year)}

\item{year}{by default "UNKN" - the format of unknown year}

\item{month}{by default "UNK" - the format of unknown month}

\item{day}{by default "UN" - the format of unknown day}
}
\value{
A logical vector
}
\description{
Check if a vector contains a complete date
}
\author{
Lukasz Andrzejewski
}
