% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_date.R
\name{check_if_year_month_entered}
\alias{check_if_year_month_entered}
\title{Check if a Vector Contains a Year and Month}
\usage{
check_if_year_month_entered(data_frame, column_name, separator = "-")
}
\arguments{
\item{data_frame}{data frame}

\item{column_name}{name of column that keeps dates to be imputed}

\item{separator}{by default "-" it is a day-month-year separator, for example "2024-10-21" has "-" separator}
}
\value{
A logical vector
}
\description{
This function determines whether the elements of a vector contain a **year** and **month**
in the specified order. It returns a logical vector indicating which elements meet this criterion.
}
\author{
Lukasz Andrzejewski
}
