% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_to_iso.R
\name{choose_mdy_format}
\alias{choose_mdy_format}
\title{Get TRUE if date format is mdy}
\usage{
choose_mdy_format(df_column)
}
\arguments{
\item{df_column}{data frame date column or vector with dates}
}
\value{
logical vector, TRUE if most probable date format is MDY
}
\description{
Get TRUE if date format is mdy
}
\author{
Lukasz Andrzejewski
}
