% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_to_iso.R
\name{choose_ydm_format}
\alias{choose_ydm_format}
\title{Get TRUE if date format is ydm}
\usage{
choose_ydm_format(df_column)
}
\arguments{
\item{df_column}{data frame date column or vector with dates}
}
\value{
logical vector, TRUE if most probable date format is YDM
}
\description{
Get TRUE if date format is ydm
}
\author{
Lukasz Andrzejewski
}
