% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_to_iso.R
\name{choose_ymd_format}
\alias{choose_ymd_format}
\title{Get TRUE if date format is ymd}
\usage{
choose_ymd_format(df_column)
}
\arguments{
\item{df_column}{data frame date column or vector with dates}
}
\value{
logical vector, TRUE if most probable date format is YMD
}
\description{
Get TRUE if date format is ymd
}
\author{
Lukasz Andrzejewski
}
