% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfiso.R
\name{dfiso}
\alias{dfiso}
\title{Function recognize date variables and modify them to ISO standard ("International Organization for Standardization")}
\usage{
dfiso(df)
}
\arguments{
\item{df}{data frame or variable/s, for example data.frame(date=c("12-Mar-2021","01-Jan-2023"))}
}
\value{
dates formatted to ISO standard (yyyy-mm-dd)
}
\description{
Function recognize date variables and modify them to ISO standard ("International Organization for Standardization")
}
\examples{
# data frame with different formatted dates
dfiso(data.frame(date1=c("13-02-2022","13/Feb/2022","13-Feb-2022")))

}
\author{
Lukasz Andrzejewski
}
