% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_to_iso.R
\name{find_dym_date_format}
\alias{find_dym_date_format}
\title{Find DYM dates only}
\usage{
find_dym_date_format(df_column)
}
\arguments{
\item{df_column}{data frame date column or vector with dates}
}
\value{
logical vector, TRUE if date format is DYM
}
\description{
Find DYM dates only
}
\author{
Lukasz Andrzejewski
}
