% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_to_iso.R
\name{find_only_dates}
\alias{find_only_dates}
\title{Return TRUE if data frame column or vector contains date}
\usage{
find_only_dates(df_column)
}
\arguments{
\item{df_column}{data frame date column or vector with dates}
}
\value{
logical vector, return TRUE if number of characters is higher than 5, contains digits and special characters
or month names
}
\description{
Return TRUE if data frame column or vector contains date
}
\author{
Lukasz Andrzejewski
}
