% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_to_iso.R
\name{find_unknow_date}
\alias{find_unknow_date}
\title{Find Unknown date, defined as UN or UNK}
\usage{
find_unknow_date(df_column)
}
\arguments{
\item{df_column}{data frame date column or vector with dates}
}
\value{
logical vector, TRUE if "un" character is found but not "jun"
}
\description{
Find Unknown date, defined as UN or UNK
}
\author{
Lukasz Andrzejewski
}
