% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_to_iso.R
\name{find_ydm_date_format}
\alias{find_ydm_date_format}
\title{Find YDM dates only}
\usage{
find_ydm_date_format(df_column)
}
\arguments{
\item{df_column}{data frame date column or vector with dates}
}
\value{
logical vector, TRUE if date format is YDM
}
\description{
Find YDM dates only
}
\author{
Lukasz Andrzejewski
}
