% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_to_iso.R
\name{find_ymd_date_format}
\alias{find_ymd_date_format}
\title{Find YMD dates only}
\usage{
find_ymd_date_format(df_column)
}
\arguments{
\item{df_column}{data frame date column or vector with dates}
}
\value{
logical vector, TRUE if date format is YMD
}
\description{
Find YMD dates only
}
\author{
Lukasz Andrzejewski
}
