% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_to_iso.R
\name{get_abbreviated_month_name}
\alias{get_abbreviated_month_name}
\title{Replace full month name by abbreviated month name}
\usage{
get_abbreviated_month_name(df_column)
}
\arguments{
\item{df_column}{data frame date column or vector with dates}
}
\value{
vector, if any full length month name, then replace by abbreviated month name
}
\description{
Replace full month name by abbreviated month name
}
\author{
Lukasz Andrzejewski
}
