% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_to_iso.R
\name{get_max_score_within_data_formats}
\alias{get_max_score_within_data_formats}
\title{Score each of date format ymd, ydm, dmy, dym, mdy, myd and return only the highest score}
\usage{
get_max_score_within_data_formats(df_column)
}
\arguments{
\item{df_column}{data frame date column or vector with dates}
}
\value{
return score of most probable date format
}
\description{
Score each of date format ymd, ydm, dmy, dym, mdy, myd and return only the highest score
}
\author{
Lukasz Andrzejewski
}
