% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_to_iso.R
\name{get_months}
\alias{get_months}
\title{List month names: full names and abbreviated  names in lower case}
\usage{
get_months()
}
\value{
full names and abbreviations of months
}
\description{
List month names: full names and abbreviated  names in lower case
}
\author{
Lukasz Andrzejewski
}
