% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_to_iso.R
\name{get_months_sep_by_vertical_bar}
\alias{get_months_sep_by_vertical_bar}
\title{Get vector with full and abbreviated name of months separated by vertical bar}
\usage{
get_months_sep_by_vertical_bar()
}
\value{
full names and abbreviations of months separated by vertical bar
}
\description{
Get vector with full and abbreviated name of months separated by vertical bar
}
\author{
Lukasz Andrzejewski
}
