% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_to_iso.R
\name{get_number_of_symbols_in_string}
\alias{get_number_of_symbols_in_string}
\title{Function to find number of symbols in date}
\usage{
get_number_of_symbols_in_string(df_column, symbol = "T")
}
\arguments{
\item{df_column}{data frame date column or vector with dates}

\item{symbol}{symbol that needs to be found, by default "T"}
}
\value{
number of found symbols
}
\description{
Function to find number of symbols in date
}
\author{
Lukasz Andrzejewski
}
