% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_to_iso.R
\name{get_up_to_12_char}
\alias{get_up_to_12_char}
\title{function return observations with up to 12 characters}
\usage{
get_up_to_12_char(df_column)
}
\arguments{
\item{df_column}{data frame column or vector to extract observarions up to 12 characters}
}
\value{
return up to 12 characters
}
\description{
function return observations with up to 12 characters
}
\author{
Lukasz Andrzejewski
}
