% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_to_iso.R
\name{has_dash_or_slash_or_white_space_characters_or_months_separated_by_vertical_bar}
\alias{has_dash_or_slash_or_white_space_characters_or_months_separated_by_vertical_bar}
\title{Function return special characters and months separated by vertical bars}
\usage{
has_dash_or_slash_or_white_space_characters_or_months_separated_by_vertical_bar(
  
)
}
\value{
special characters and months: "-|\\/|\\w+\\s+|january|february|march|april|may|june|july|august|september|october|november|december|jan|feb|mar|apr|may|jun|jul|aug|sep|oct|nov|dec"
}
\description{
Function return special characters and months separated by vertical bars
}
\author{
Lukasz Andrzejewski
}
