% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_to_iso.R
\name{has_dash_or_slash_or_white_space_characters_separated_by_vertical_bar}
\alias{has_dash_or_slash_or_white_space_characters_separated_by_vertical_bar}
\title{Function return special characters separated by vertical bars}
\usage{
has_dash_or_slash_or_white_space_characters_separated_by_vertical_bar(
  special_characters = c("-", "\\\\/", "\\\\w+\\\\s+")
)
}
\arguments{
\item{special_characters}{by default dash, slash, white space characters}
}
\value{
special characters: "-|\\/|\\w+\\s+"
}
\description{
Function return special characters separated by vertical bars
}
\author{
Lukasz Andrzejewski
}
