% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_to_iso.R
\name{prepare_date}
\alias{prepare_date}
\title{Additional step for YMD date type}
\usage{
prepare_date(df_column)
}
\arguments{
\item{df_column}{data frame date column or vector with dates}
}
\value{
output up to 12 characters, remove whitespace from start and end of string, keep characters from
the left site of letter "T"
}
\description{
Additional step for YMD date type
}
\author{
Lukasz Andrzejewski
}
