% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_to_iso.R
\name{remove_unnecessary_part_of_date}
\alias{remove_unnecessary_part_of_date}
\title{Get substring of date to eliminate unnecessary part}
\usage{
remove_unnecessary_part_of_date(df_column, symbol = "T")
}
\arguments{
\item{df_column}{date column or vector with dates}

\item{symbol}{symbol that needs to be found, by default "T"}
}
\value{
substring of date from position 1 to position where last "symbol" is located
}
\description{
Get substring of date to eliminate unnecessary part
}
\author{
Lukasz Andrzejewski
}
