% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viso.R
\name{viso}
\alias{viso}
\title{transform date vector to date vector in ISO standard ("International Organization for Standardization")}
\usage{
viso(df_column)
}
\arguments{
\item{df_column}{vector or string}
}
\value{
dates formatted to ISO standard (yyyy-mm-dd)
}
\description{
transform date vector to date vector in ISO standard ("International Organization for Standardization")
}
\examples{
#day month year vector
viso(c("12Mar2022","21Feb2022"))

#day month year vector in different formats
viso(c("12Mar2022","21-02-2022"))

#month year day vector
viso(c("Mar-2022-12","Feb-2022-21"))
}
\author{
Lukasz Andrzejewski
}
