% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_to_iso.R
\name{remove_no_date_characters}
\alias{remove_no_date_characters}
\title{Function remove symbols}
\usage{
remove_no_date_characters(df_column, symbols = "[;:+]")
}
\arguments{
\item{df_column}{data frame column or vector from which symbols need to be removed}

\item{symbols}{by default ; : +}
}
\value{
by default delete semicolon, colon and plus sign from vector or data frame
}
\description{
Function remove symbols
}
\author{
Lukasz Andrzejewski
}
